requirements.txt
anthropic>=0.25.0
beautifulsoup4>=4.12.0
gspread>=6.0.0
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.120.0
requests>=2.31.0
lxml>=5.1.0
playwright
python-dotenv